+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Nothing unreleased.
 
-## [1.1.300] - 2024-06-07
+## [1.1.300rc5] - 2024-06-11
 
 ### Added